XAdmin UI kebab case
