+++ packages/local/XAdmin/ui/src/components/forms/smart-connect-form.tsx
@@ -1,10 +1,10 @@
-import { queryKeys } from "@/lib/queryKeys";
+import { queryKeys } from "@/lib/query-keys";
 
 import { queryClient } from "@shared/lib/query-client";
 
 import { Schema, UrlForm } from "../../components/forms/url";
 import { useToast } from "../../components/ui/use-toast";
-import { useConnect } from "../../hooks/useConnect";
+import { useConnect } from "../../hooks/use-connect";
 
 interface Props {
     onConnection: () => void;
