+++ packages/local/XAdmin/ui/src/components/prev-next-buttons.tsx
