+++ packages/local/XAdmin/ui/src/hooks/use-import-account.ts
@@ -1,7 +1,7 @@
 import { useMutation } from "@tanstack/react-query";
 import z from "zod";
 
-import { queryKeys } from "@/lib/queryKeys";
+import { queryKeys } from "@/lib/query-keys";
 
 const zCredentialSchema = z.object({
     account: z.string(),
