+++ packages/local/XAdmin/ui/src/hooks/use-performance.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
 
-import { chain } from "@/lib/chainEndpoints";
+import { chain } from "@/lib/chain-endpoints";
 
 export const usePerformance = () =>
     useQuery({
