+++ packages/local/XAdmin/ui/src/hooks/use-statuses.ts
@@ -1,8 +1,8 @@
 import { useQuery } from "@tanstack/react-query";
 import { z } from "zod";
 
-import { chain } from "@/lib/chainEndpoints";
-import { queryKeys } from "@/lib/queryKeys";
+import { chain } from "@/lib/chain-endpoints";
+import { queryKeys } from "@/lib/query-keys";
 
 export const useStatuses = () =>
     useQuery<string[], string>({
