+++ packages/local/XAdmin/ui/src/hooks/use-templated-packages.ts
