--- packages/local/XAdmin/ui/src/lib/chainEndpoints/index.ts
@@ -1,3 +1,0 @@
-export * from "./packageExtraction";
-export * from "./packageApi";
-export * from "./installFiles";
