+++ packages/local/XAdmin/ui/src/lib/get-required-packages.ts
