+++ packages/local/XAdmin/ui/src/lib/recursive-fetch.ts
