+++ packages/local/XAdmin/ui/src/lib/split-version.ts
