+++ packages/local/XAdmin/ui/src/pages/join-page.tsx
@@ -4,7 +4,7 @@ import { useToast } from "@/components/ui/use-toast";
 
 import { Schema, UrlForm } from "@/components/forms/url";
 
-import { useConnect } from "../hooks/useConnect";
+import { useConnect } from "../hooks/use-connect";
 import { SetupWrapper } from "./setup-wrapper";
 
 export const JoinPage = () => {
